reverting to past version where bugs were not introduced from trying to implement location & camera

+++ app/src/main/java/com/example/runqr/Location.java
@@ -27,7 +27,7 @@ public class Location{
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
@@ -36,7 +36,7 @@ public class Location{
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
